+++ tests/test_parallels.py
@@ -8,10 +8,6 @@
 # Imports
 ##############################################################################
 
-# enable some python3 compatibility options:
-# (unicode_literals not compatible with python2 uuid module)
-from __future__ import absolute_import, print_function
-
 import py_trees
 import py_trees.console as console
 
@@ -28,9 +24,7 @@ logger = py_trees.logging.Logger("Nosetest")
 
 
 def test_parallel_failure():
-    print(console.bold + "\n****************************************************************************************" + console.reset)
-    print(console.bold + "* Parallel Failure" + console.reset)
-    print(console.bold + "****************************************************************************************\n" + console.reset)
+    console.banner("Parallel Failure")
     root = py_trees.composites.Parallel("Parallel")
     failure = py_trees.behaviours.Failure("Failure")
     success = py_trees.behaviours.Success("Success")
@@ -41,18 +35,16 @@ def test_parallel_failure():
     py_trees.tests.tick_tree(root, visitor, 1, 1)
 
     print("\n--------- Assertions ---------\n")
-    print("root.status == py_trees.Status.FAILURE")
-    assert(root.status == py_trees.Status.FAILURE)
-    print("failure.status == py_trees.Status.FAILURE")
-    assert(failure.status == py_trees.Status.FAILURE)
-    print("success.status == py_trees.Status.SUCCESS")
-    assert(success.status == py_trees.Status.SUCCESS)
+    print("root.status == py_trees.common.Status.FAILURE")
+    assert(root.status == py_trees.common.Status.FAILURE)
+    print("failure.status == py_trees.common.Status.FAILURE")
+    assert(failure.status == py_trees.common.Status.FAILURE)
+    print("success.status == py_trees.common.Status.SUCCESS")
+    assert(success.status == py_trees.common.Status.SUCCESS)
 
 
 def test_parallel_success():
-    print(console.bold + "\n****************************************************************************************" + console.reset)
-    print(console.bold + "* Parallel Success" + console.reset)
-    print(console.bold + "****************************************************************************************\n" + console.reset)
+    console.banner("Parallel Success")
     root = py_trees.composites.Parallel("Parallel")
     success1 = py_trees.behaviours.Success("Success1")
     success2 = py_trees.behaviours.Success("Success2")
@@ -63,20 +55,24 @@ def test_parallel_success():
     py_trees.tests.tick_tree(root, visitor, 1, 1)
 
     print("\n--------- Assertions ---------\n")
-    print("root.status == py_trees.Status.SUCCESS")
-    assert(root.status == py_trees.Status.SUCCESS)
-    print("success1.status == py_trees.Status.SUCCESS")
-    assert(success1.status == py_trees.Status.SUCCESS)
-    print("success2.status == py_trees.Status.SUCCESS")
-    assert(success2.status == py_trees.Status.SUCCESS)
+    print("root.status == py_trees.common.Status.SUCCESS")
+    assert(root.status == py_trees.common.Status.SUCCESS)
+    print("success1.status == py_trees.common.Status.SUCCESS")
+    assert(success1.status == py_trees.common.Status.SUCCESS)
+    print("success2.status == py_trees.common.Status.SUCCESS")
+    assert(success2.status == py_trees.common.Status.SUCCESS)
 
 
 def test_parallel_running():
-    print(console.bold + "\n****************************************************************************************" + console.reset)
-    print(console.bold + "* Parallel Running" + console.reset)
-    print(console.bold + "****************************************************************************************\n" + console.reset)
-    root = py_trees.composites.Parallel("Parallel", policy=py_trees.common.ParallelPolicy.SUCCESS_ON_ALL)
-    success_after_1 = py_trees.behaviours.Count(name="SuccessAfter1", fail_until=0, running_until=1, success_until=20)
+    console.banner("Parallel Running")
+    root = py_trees.composites.Parallel(
+        "Parallel", policy=py_trees.common.ParallelPolicy.SUCCESS_ON_ALL)
+    success_after_1 = py_trees.behaviours.Count(
+        name="SuccessAfter1",
+        fail_until=0,
+        running_until=1,
+        success_until=20,
+        reset=False)
 
     running = py_trees.behaviours.Running("Running")
     success_every_other = py_trees.behaviours.SuccessEveryN("SuccessEveryOther", 2)
@@ -88,34 +84,33 @@ def test_parallel_running():
     py_trees.tests.tick_tree(root, visitor, 1, 1)
 
     print("\n--------- Assertions ---------\n")
-    print("root.status == py_trees.Status.FAILURE")
-    assert(root.status == py_trees.Status.FAILURE)
-    print("success_after_1.status == py_trees.Status.FAILURE")
-    assert(success_after_1.status == py_trees.Status.FAILURE)
-    print("running.status == py_trees.Status.FAILURE")
-    assert(running.status == py_trees.Status.FAILURE)
-    print("success_every_other.status == py_trees.Status.FAILURE")
-    assert(success_every_other.status == py_trees.Status.FAILURE)
+    print("root.status == py_trees.common.Status.FAILURE")
+    assert(root.status == py_trees.common.Status.FAILURE)
+    print("success_after_1.status == py_trees.common.Status.INVALID")
+    assert(success_after_1.status == py_trees.common.Status.INVALID)
+    print("running.status == py_trees.common.Status.INVALID")
+    assert(running.status == py_trees.common.Status.INVALID)
+    print("success_every_other.status == py_trees.common.Status.FAILURE")
+    assert(success_every_other.status == py_trees.common.Status.FAILURE)
 
     py_trees.tests.tick_tree(root, visitor, 2, 2)
 
     print("\n--------- Assertions ---------\n")
-    print("root.status == py_trees.Status.RUNNING")
-    assert(root.status == py_trees.Status.RUNNING)
-    print("success_after_1.status == py_trees.Status.SUCCESS")
-    assert(success_after_1.status == py_trees.Status.SUCCESS)
-    print("running.status == py_trees.Status.RUNNING")
-    assert(running.status == py_trees.Status.RUNNING)
-    print("success_every_other.status == py_trees.Status.SUCCESS")
-    assert(success_every_other.status == py_trees.Status.SUCCESS)
+    print("root.status == py_trees.common.Status.RUNNING")
+    assert(root.status == py_trees.common.Status.RUNNING)
+    print("success_after_1.status == py_trees.common.Status.SUCCESS")
+    assert(success_after_1.status == py_trees.common.Status.SUCCESS)
+    print("running.status == py_trees.common.Status.RUNNING")
+    assert(running.status == py_trees.common.Status.RUNNING)
+    print("success_every_other.status == py_trees.common.Status.SUCCESS")
+    assert(success_every_other.status == py_trees.common.Status.SUCCESS)
 
 
 def test_parallel_success_on_one():
-    print(console.bold + "\n****************************************************************************************" + console.reset)
-    print(console.bold + "* Parallel Success on One" + console.reset)
-    print(console.bold + "****************************************************************************************\n" + console.reset)
+    console.banner("Parallel Success on One")
     print("")
-    root = py_trees.composites.Parallel(name="Parallel", policy=py_trees.common.ParallelPolicy.SUCCESS_ON_ONE)
+    root = py_trees.composites.Parallel(
+        name="Parallel", policy=py_trees.common.ParallelPolicy.SUCCESS_ON_ONE)
     running1 = py_trees.behaviours.Running("Running1")
     success = py_trees.behaviours.Success("Success")
     running2 = py_trees.behaviours.Running("Running2")
@@ -128,12 +123,11 @@ def test_parallel_success_on_one():
 
     print("\n--------- Assertions ---------\n")
     print("All children get switched to success if one goes to success.")
-    print("root.status == py_trees.Status.SUCCESS")
-    assert(root.status == py_trees.Status.SUCCESS)
-    print("running1.status == py_trees.Status.SUCCESS")
-    assert(running1.status == py_trees.Status.SUCCESS)
-    print("success.status == py_trees.Status.SUCCESS")
-    assert(success.status == py_trees.Status.SUCCESS)
-    print("running2.status == py_trees.Status.SUCCESS")
-    assert(running2.status == py_trees.Status.SUCCESS)
-
+    print("root.status == py_trees.common.Status.SUCCESS")
+    assert(root.status == py_trees.common.Status.SUCCESS)
+    print("running1.status == py_trees.common.Status.INVALID")
+    assert(running1.status == py_trees.common.Status.INVALID)
+    print("success.status == py_trees.common.Status.SUCCESS")
+    assert(success.status == py_trees.common.Status.SUCCESS)
+    print("running2.status == py_trees.common.Status.INVALID")
+    assert(running2.status == py_trees.common.Status.INVALID)